build: fix build issues after upgrading dependencies
