+++ package.json
@@ -2,8 +2,10 @@
   "name": "rollup-plugin-summary",
   "version": "0.0.0-semantically-released",
   "description": "A rollup plugin that summarizes the build",
-  "main": "src/index.ts",
-  "typings": "index.d.ts",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/es/index.js",
+  "types": "./index.d.ts",
+  "typings": "./index.d.ts",
   "repository": "https://github.com/yousifalraheem/rollup-plugin-summary.git",
   "homepage": "https://yousifalraheem.github.io/rollup-plugin-summary/",
   "bugs": {
@@ -21,9 +23,11 @@
     "table"
   ],
   "files": [
-    "src/index.ts",
-    "index.js.map",
-    "src/index.d.ts"
+    "dist",
+    "!dist/**/*.map",
+    "types",
+    "README.md",
+    "LICENSE"
   ],
   "scripts": {
     "prebuild": "rm -rf dist/*",
@@ -53,10 +57,10 @@
     "@babel/preset-typescript": "^7.18.6",
     "@commitlint/cli": "^17.2.0",
     "@commitlint/config-conventional": "^17.2.0",
-    "@rollup/plugin-commonjs": "^22.0.2",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-typescript": "^8.5.0",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-json": "^5.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^9.0.2",
     "@types/jest": "^29.2.2",
     "@types/node": "^18.11.9",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
@@ -74,16 +78,15 @@
     "lint-staged": "^13.0.3",
     "prettier": "^2.7.1",
     "regenerator-runtime": "^0.13.10",
-    "rollup": "^2.79.1",
-    "rollup-plugin-node-externals": "^4.1.1",
+    "rollup": "^3.0.0",
+    "rollup-plugin-node-externals": "^5.0.2",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-terser": "^7.0.2",
     "semantic-release": "^19.0.5",
     "ts-jest": "^29.0.3",
     "typescript": "^4.8.4"
   },
   "peerDependencies": {
-    "rollup": "^2.0.0"
+    "rollup": "^2.68.0||^3.0.0"
   },
   "config": {
     "commitizen": {
