+++ test-project/rollup.config.js
@@ -2,8 +2,7 @@ import commonjs from "@rollup/plugin-commonjs";
 import resolve from "@rollup/plugin-node-resolve";
 import externals from "rollup-plugin-node-externals";
 import { defineConfig } from "rollup";
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { summary } = require("../dist");
+import summary from "rollup-plugin-summary";
 
 const configs = defineConfig({
   input: "index.js",
